pyproject.toml
[tool.poetry]
name = "scrape2"
version = "0.1.0"
description = ""
authors = ["Jai Kang <1807929+jdkang@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
boto3 = "^1.18.48"
botocore = "^1.21.48"
click = "^8.0.1"
click-help-colors = "^0.9.1"
google-api-python-client = "^2.22.0"
selenium = "^3.141.0"
sty = "^1.0.0-rc.2"
newspaper3k = "^0.2.8"
pocket = "^0.3.6"
google-auth-oauthlib = "^0.4.6"
schedule = "^1.1.0"
loguru = "^0.5.3"
Jinja2 = "^3.0.2"
slack-bolt = "^1.9.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.9b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
